# doc/api.yaml
openapi: 3.0.0
paths:
...
  /user:
    get:
      operationId: UserController_findMany
      summary: Get all users
      description: Gets all users
      parameters: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401": &a2
          description: Access token is missing or invalid
      tags: &a3
        - Users
      security: &a4
        - access-token: []
    post:
      operationId: UserController_create
      summary: Create new user
      description: Creates a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserDto"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: body does not contain required fields
        "401": *a2
      tags: *a3
      security: *a4
  /user/{id}:
    get:
      operationId: UserController_findOne
      summary: Get user by id
      description: Gets user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "200":
          description: User data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: id is invalid (not UUID)
        "401": *a2
        "404":
          description: User not found
      tags: *a3
      security: *a4
    put:
      operationId: UserController_update
      summary: Update user password by id
      description: Updates user password by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePasswordDto"
      responses:
        "200":
          description: Updated user password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedUserDto"
        "400":
          description: id is invalid (not UUID)
        "401": *a2
        "403":
          description: Invalid oldPassword
        "404":
          description: User not found
      tags: *a3
      security: *a4
    delete:
      operationId: UserController_delete
      summary: Delete user by id
      description: Deletes user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        "204":
          description: User deleted
        "400":
          description: id is invalid (not UUID)
        "401": *a2
        "404":
          description: User not found
      tags: *a3
      security: *a4
...
info:
  title: Home Library Service API
  description: Home music library service
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: Authorization
      in: header
  schemas:
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: cat_lover
          description: User login
        password:
          type: string
          example: "12345"
          description: User password
      required:
        - login
        - password
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: f0f9b1b7-8b2b-4b6b-8b1b-8b2b8b1b7b2b
          description: User ID
        login:
          type: string
          example: cat_lover
          description: User login
        version:
          type: number
          example: 1
          description: User version
        createdAt:
          type: number
          example: 1710084738234
          description: User creation time
        updatedAt:
          type: number
          example: 1710084738234
          description: User update time
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: "12345"
          description: User old password
        newPassword:
          type: string
          example: "123456789"
          description: User new password
      required:
        - oldPassword
        - newPassword
    UpdatedUserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: f0f9b1b7-8b2b-4b6b-8b1b-8b2b8b1b7b2b
          description: User ID
        login:
          type: string
          example: cat_lover
          description: User login
        createdAt:
          type: number
          example: 1710084738234
          description: User creation time
        version:
          type: number
          example: 2
          description: User version
        updatedAt:
          type: number
          example: 1710094738234
          description: User update time
      required:
        - id
        - login
        - createdAt
        - version
        - updatedAt
...